fix: Fix panic in state recovery tests
- Add nil check in recoverServicesForCluster for storage availability
- Create shared test_helpers.go with MockClusterManager definition
- Remove duplicate MockClusterManager from server_shutdown_test.go
- Fix import issues in test files

The panic was caused by the test not properly initializing the storage's
ServiceStore, which returned nil when accessed.
